Implemented cache before removing.

+++ v2/src/__use_cases__/crop/remove-crop.js
@@ -16,11 +16,39 @@ module.exports = makeDeleteCrop = ({ dirtyCache, cropDb }) => {
         message: "Cannot be deleted, under processing.",
       };
     }
+
+    if (!dirtyCache.add({ id })) {
+      return {
+        deletedCount: 0,
+        serverError: 1,
+        message:
+          "Internal Server Error. Crop id cannot be added to the dirty cache.",
+      };
+    }
+
     const existing = await cropDb.findById({ id });
     if (!existing) {
+      if (!dirtyCache.remove({ id })) {
+        return {
+          deletedCount: 0,
+          serverError: 1,
+          message:
+            "Internal Server Error. Crop id cannot be removed from the dirty cache.",
+        };
+      }
       return deleteNothing();
     }
+
     await cropDb.removeById({ id });
+    if (!dirtyCache.remove({ id })) {
+      return {
+        deletedCount: 0,
+        serverError: 1,
+        message:
+          "Internal Server Error. Crop id cannot be removed from the dirty cache.",
+      };
+    }
+
     return {
       deleteCount: 1,
       message: "Crop deleted Successfully.",
